rinad: ipcm: move catalog into ipcm compilation unit
If needed, catalog will be moved back in its own unit in future
patches.
