+++ rinad/src/ipcm/ipcm.h
@@ -239,6 +239,38 @@ public:
 	virtual ~SyscallTransState(){};
 };
 
+// Data structure that represent a plugin in the catalog
+struct CatalogPlugin {
+	// The name of the plugin
+	std::string name;
+
+	// Is the plugin already loaded ?
+	bool loaded;
+};
+
+struct CatalogPsInfo: public rina::PsInfo {
+	// Back-reference to the plugin that published this
+	// policy-set
+	std::map<std::string, CatalogPlugin>::iterator plugin;
+
+	// Is the policy-set alread loaded (i.e. is the associated
+	// plugin already loaded ?
+	bool loaded;
+
+	CatalogPsInfo() : PsInfo() { }
+	CatalogPsInfo(const std::string& n, const std::string& c,
+	              const std::string& v);
+};
+
+class Catalog {
+public:
+	Catalog();
+
+private:
+	std::map<std::string, CatalogPsInfo> policy_sets;
+	std::map<std::string, CatalogPlugin> plugins;
+};
+
 //
 // @brief The IPCManager class is in charge of managing the IPC processes
 // life-cycle.
@@ -829,6 +861,9 @@ private:
 
 	//Stop condition
 	rina::ConditionVariable stop_cond;
+
+	//Catalog
+	Catalog catalog;
 
 	//Trampoline for the pthread_create
 	static void* io_loop_trampoline(void* param);
